+++ formatters/date.js
@@ -1,4 +1,4 @@
-var moment = require('moment');
+var dayjs = require('dayjs');
 
 
 /**
@@ -6,19 +6,19 @@ var moment = require('moment');
  *
  * Since 1.2.0, by default, it considers the input format is "ISO 8601"
  *
- * @exampleContext {"lang":"en"}
+ * @exampleContext {"lang":"en", "timezone":"Europe/Paris"}
  * @example ["20160131", "L"]
  * @example ["20160131", "LL"]
  * @example ["20160131", "LLLL"]
  * @example ["20160131", "dddd"]
  *
- * @exampleContext {"lang":"fr"}
+ * @exampleContext {"lang":"fr", "timezone":"Europe/Paris"}
  * @example ["2017-05-10T15:57:23.769561+03:00", "LLLL"]
  * @example ["2017-05-10 15:57:23.769561+03:00", "LLLL"]
  * @example ["20160131", "LLLL"]
  * @example ["20160131", "dddd"]
  *
- * @exampleContext {"lang":"fr"}
+ * @exampleContext {"lang":"fr", "timezone":"Europe/Paris"}
  * @example ["20160131", "dddd", "YYYYMMDD"]
  * @example [1410715640, "LLLL", "X" ]
  *
@@ -29,11 +29,7 @@ var moment = require('moment');
  */
 function formatD (d, patternOut, patternIn) {
   if (d !== null && typeof d !== 'undefined') {
-    moment.locale(this.lang);
-    if (patternIn) {
-      return moment(d + '', patternIn).format(patternOut);
-    }
-    return moment(d + '').format(patternOut);
+    return fromMomentToDayJS(d, patternIn).tz(this.timezone).locale(this.lang).format(patternOut);
   }
   return d;
 }
@@ -44,14 +40,14 @@ function formatD (d, patternOut, patternIn) {
  *
  * @deprecated
  *
- * @exampleContext {"lang":"en"}
+ * @exampleContext {"lang":"en", "timezone":"Europe/Paris"}
  * @example ["20160131", "YYYYMMDD", "L"]
  * @example ["20160131", "YYYYMMDD", "LL"]
  * @example ["20160131", "YYYYMMDD", "LLLL"]
  * @example ["20160131", "YYYYMMDD", "dddd"]
  * @example [1410715640, "X", "LLLL"]
  *
- * @exampleContext {"lang":"fr"}
+ * @exampleContext {"lang":"fr", "timezone":"Europe/Paris"}
  * @example ["20160131", "YYYYMMDD", "LLLL"]
  * @example ["20160131", "YYYYMMDD", "dddd"]
  *
@@ -61,13 +57,28 @@ function formatD (d, patternOut, patternIn) {
  * @return {String}            return formatted date
  */
 function convDate (d, patternIn, patternOut) {
-  if (d !== null && typeof d !== 'undefined') {
-    moment.locale(this.lang);
-    return moment(d + '', patternIn).format(patternOut);
-  }
-  return d;
+  return formatD(d, patternOut, patternIn);
 }
 
+/**
+ * Convert old MomentJS format to DayJS format
+ *
+ * @param      {string}  d          not undefined/null date
+ * @param      {string}  patternIn  The pattern
+ * @return     {Object}             dayjs
+ */
+function fromMomentToDayJS (d, patternIn) {
+  if (!patternIn) {
+    return dayjs(d + '');
+  }
+  if (patternIn === 'X') {
+    return dayjs.unix(d);
+  }
+  if (patternIn === 'x') {
+    return dayjs(d);
+  }
+  return dayjs(d + '', patternIn);
+}